requirements.txt
Django>=2.0.2
django-debug-toolbar>=1.9.1
django-phonenumber-field>=2.0.0
django-solo>=1.1.3
djangorestframework>=3.7.7
